Tarefa1_2022li1g001

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

Funcao que verifica se um mapa é ou não válido conforme as restrições

obstaculoValido :: Mapa -> Bool

Funcoes de restricao para mapaValido

Verifica se exitem obstaculos em terrenos impróprios.

rioContiguoValido :: Mapa -> Bool

Verifica se rios contiguos possuem direcoes opostas

troncoValido :: Mapa -> Bool

Verifica se o comprimento do Tronco e valido (<=5)

carroValido :: Mapa -> Bool

Verifica se o comprimento do Carro e valido (<=3)

nenhumValido :: Mapa -> Bool

Verifica se existe pelo menos um espaço livre numa linha

larguraValida :: Mapa -> Bool

Verifica se que o comprimento da lista de obstaculos é corresponde à largura do mapa

terrenosContiguos :: Mapa -> Bool

Verifca que não existem mais de 4 rios, nem 5 estradas ou relvas contiguamente

comprimentoValido :: Eq a => [a] -> a -> Int

Funcoes auxiliares

Funcao complementar da maxSeq

maxSeq :: Eq a => [a] -> a -> Int -> Int

Funcao que devolve o tamanho da maior sequencia de um elemento numa lista

count :: Eq a => [a] -> a -> Int

listaTerrenos :: [(Terreno, [Obstaculo])] -> [Terreno]

Funcao que recebe a lista do mapa e devolve uma lista apenas com os tipos de terreno

remVel :: [Terreno] -> [String]

Funcao que recebe uma lista de terrenos e devolve uma lista com as strings correspondentes, util para remover o fator velocidade

listaObstaculos :: [(Terreno, [Obstaculo])] -> [[Obstaculo]]

Funcao que recebe a lista do mapa e devolve uma lista com as listas de obstaculos