Tarefa2_2022li1g001

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

estendeMapa :: Mapa -> Int -> Mapa

Função que estende o Mapa

proximosTerrenosValidos :: Mapa -> [Terreno]

Funcao que dá a lista de terrenos válidos para a proxima linha do dado mapa

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

Funcao que dá a lista de obstaculos válidos a serem adicionados na linha

terrenosIguais :: [(Terreno, [Obstaculo])] -> Int

Funcao que devolve o numero de terrenos iguais contíguos

strTerreno :: (Terreno, [Obstaculo]) -> String

Funcao que recebe um terreno e devolve a string correspondente, util para remover velocidades

obstaculosIguais :: [Obstaculo] -> Int -> Int

Funcao que conta o numero de obstaculos iguais no final da lista de obstaculos

listaAleatorios :: Int -> Int -> [Int]

Funcao que gera lista de inteiros psuedo-aleatorios de len elementos

escolherTerreno :: [Terreno] -> Int -> Int -> Int -> Terreno

Funcao que escolhe pseudo-aleatoriamente um dos terrenos validos para a proxima linha do mapa

gerarObstaculos :: Int -> (Terreno, [Obstaculo]) -> Int -> [Obstaculo]

Funcao que gera pseudo-aleatoriamente a lista de obstaculos validos para a linha do mapa

escolherObstaculos :: Int -> (Terreno, [Obstaculo]) -> Int -> Obstaculo

Funcao que escolhe pseudo-aleatoriamente um elemento da lista de obstaculos para um dado terreno