Tarefa3_2022li1g001

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

animaJogo :: Jogo -> Jogada -> Jogo

Funcao que movimenta os obstáculos (de acordo com a velocidade do terreno em que se encontram) e o personagem, de acordo com a jogada dada

moveMapa :: [(Terreno, [Obstaculo])] -> [(Terreno, [Obstaculo])]

Move todas as listas de obstaculos dos terrenos Rio e Estrada

moveAuxDir :: Int -> [Obstaculo] -> [Obstaculo]

Move a lista n elementos para a Direita

moveAuxEsq :: Int -> [Obstaculo] -> [Obstaculo]

Move a lista n elementos para a Esquerda

eTronco :: Jogador -> Mapa -> Int -> Jogador

Funcao que move jogador quando este se encontra em cima de um tronco, onde n representa o indice da linha do mapa

eArvore :: Jogo -> Jogada -> Bool

Funcao que verifica se é posssivel um certa jogada consoante as arvore ao redor do jogador, devolvendo True se houver uma arvore a obstruir a jogada